requirements.txt
flask
redis
mysql-connector==2.2.9